Add method to register scheduler per type and contract.
This resolves a bug in which listening for an event before sending the first message will lock all messages to the default scheduler rather than the one passed into the `SendMessage` method.
The API is now more explicit. Call `RegisterScheduler` to register a scheduler for each combination of message type and contract. If you don't register a scheduler, the default one is used. This ensures that `Listen` and `SendMessage` calls are always paired up properly on the same scheduler.
Closes Issue #47