Handle OnError's that can terminate internal IOs
One problem that has plagued ReactiveUI is how to handle what happens when an
IObservable OnError's. This is most common on the CanExecute or when a
ReactiveAsyncCommand fails. The error would end up completing a Subject
internally which means that things would silently stop working.

Instead, we're going to implement a ThrownExceptions IObservable on objects
where this can happen. If this IObservable doesn't have any Subscribers, it
will send the result to the "Default" subscriber,
RxApp.DefaultExceptionHandler, which blows up the UI thread with a descriptive
message.